requirements.txt
pyyaml
smart_open
